requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
requests==2.31.0